backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
google-cloud-aiplatform==1.38.0
openai==1.3.5
pinecone-client==2.2.4
httpx==0.25.2
python-multipart==0.0.6
Pillow==10.1.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
